Test : Added test for verifyCommand
